Trying to get selenium to play nice with TBB.
There is just all kinds of broken in selenium's firefox driver, and
TBB breaks a lot of its assumptions.  This still doesn't quite work.

On the up side, it appears that we don't need the server after all,
which means we don't need Java either (and there was much rejoicing).
Keeping the build goo for it for now, though.