New driver and organizational schema for URL sources.
The basic idea (which all scripts will adhere to from now on) is
that there is a top-level script "url-source" that just imports the
"url_sources" package and runs its "driver" method.  That method loads all
modules in the package whose name starts with "s_".  Each is expected to
provide an argparse subparser, and we dispatch to the appropriate module
in the usual "command subcommand args..." fashion (where "subcommand"
is the name of the "s_" module, sans prefix).  This is not yet fully
implemented, but the outline is there.

Source modules that aren't yet Python-3-compatible are renamed "ds_whatever"
instead of "s_whatever" to keep them out of the way for the time being.